requirements.txt
Pillow~=9.4.0
numpy~=1.24.1
datasets~=2.9.0
pytorch-lightning
lightning[extra]
seqeval
evaluate
transformers

--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio

